+++ tests/conftest.py
@@ -6,3 +6,6 @@ collect_ignore = []
 
 if not (v[0] >= 3 and v[1] >= 5):
     collect_ignore.append('test_async.py')
+
+if not (v[0] >= 3 and v[1] >= 7):
+    collect_ignore.append('test_trio.py')